refactor(platform): make node/bun websocket adapters internal modules
The websocket adapter modules were public libuild endpoints
(@b9g/platform-{node,bun}/websocket) only because the generated worker
imported them by that specifier. They're internal plumbing, not user
API, so expose them through the package's main entry instead:

- Rename websocket.ts -> _websocket.ts (libuild excludes _-prefixed files
  from generated endpoints).
- index.ts re-exports the adapter (attachNodeWebSocketHandler /
  createBunWebSocketServer) so the generated worker imports it from the
  main entry it already uses; the pool path now uses the static binding
  instead of a lazy import (avoids libuild's mixed static+dynamic chunk
  splitting of the same module).
- Generated-worker templates (in both platform.ts and index.ts, per the
  dual-file codegen) import the adapter from @b9g/platform-{node,bun}.

Result: the `/websocket` endpoints are gone; node/bun now expose only
`.`, `./index`, `./platform`. Cloudflare's websocket-do/pubsub stay
public — the generated entry must re-export those DO classes by name for
wrangler binding.

Validated: production build + run (e2e-direct-mode), bundling
(e2e-bundling), dev mode (develop), package unit tests, build-output
structure, and cloudflare WS regression — all green. typecheck + lint
clean.
